Fix bug: saved theme doesn't work
When save the current theme as a new theme, it needs to save the current palette
also the reason is that it alwasy regenerate the eutmux config file which also
needs palette to replace the color placeholders.

The palette used to replace color is saved in PALETTE_FILENAME variable. By
default, it points to DEFAULT_PALETTE_FILENAME which won't be used for new theme
generation and its content might be too old to adapt the new theme generation
and its color placeholer name might be invalid anymore. Therefore, it result in
failure in replacing color replaceholder for new generated eutmux config file.

Now, the solution is to save the new generated palette file to
EUTMUX_CONFIG_HOME as 'current.palette'. When saving the current theme as new
theme is requested, 'current.palette' will be copied to a new file with name
THEME_NAME.palette. Then, when request to set the target theme as the current
theme, the value PALETTE_FILENAME will points to the file THEME_NAME.palette
which can be used to replace colors in eutmux config file.