if palette filename is given, set dyanmic theme/palette as dynamic
The reason is that the palette file name is given, it's used to adapt the tmux
theme to the owner of the palette file, and the owner is usually the IDE such as
Emacs and Vim. In that case, don't override the default terminal theme whose
name is eutmux. yes, but maybe we should use 'terminal' as default theme name for terminal.